added atMail parameter
at mail sometimes gets sent to the account that the spawner runs under,
instead of the account running the Workspace Server.
set atmail to 'Y' to let at send the mail automatically, which will have
the at job number in the mail subject line.
set atmail to 'N' to send the log and output separately to &sysuserid,
which will NOT have the at job number in the mail subject line.